web-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.4
python-dotenv==1.0.0
